+++ src/migration/formatters/formatter-registry.ts
@@ -3,6 +3,7 @@ import { MigrationContext } from "../data/migration-context";
 import { Biome } from "./biome";
 import { Formatter } from "./formatter";
 import { NoFormatter } from "./no-formatter";
+import { Oxfmt } from "./oxfmt";
 import { Prettier } from "./prettier";
 
 //----------------------------------------------------------------------------------------------------------------------
@@ -10,9 +11,10 @@ import { Prettier } from "./prettier";
 //----------------------------------------------------------------------------------------------------------------------
 
 const FORMATTERS = {
-    disabled: new NoFormatter(),
-    prettier: new Prettier(),
     biome: new Biome(),
+    disabled: new NoFormatter(),
+    oxfmt: new Oxfmt(),
+    prettier: new Prettier(),
 } as const satisfies Record<NewConfig["formatter"]["value"] | OldConfig["formatter"]["value"], Formatter>;
 
 //----------------------------------------------------------------------------------------------------------------------